Move json_upload stuff into the JsonUploadMixin.
