+++ openslides_backend/action/actions/topic/json_upload.py
@@ -1,26 +1,20 @@
-from time import time
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 
 import fastjsonschema
 
 from ....models.models import Topic
 from ....permissions.permissions import Permissions
-from ....shared.interfaces.event import Event, EventType
-from ....shared.interfaces.write_request import WriteRequest
-from ....shared.patterns import fqid_from_collection_and_id
 from ....shared.schema import required_id_schema
-from ...action import Action
-from ...mixins.import_mixins import ImportStatus
+from ...mixins.import_mixins import ImportStatus, JsonUploadMixin
 from ...util.default_schema import DefaultSchema
 from ...util.register import register_action
-from ...util.typing import ActionResultElement
 from ..agenda_item.agenda_creation import agenda_creation_properties
 from .create import TopicCreate
 from .mixins import DuplicateCheckMixin
 
 
 @register_action("topic.json_upload")
-class TopicJsonUpload(DuplicateCheckMixin, Action):
+class TopicJsonUpload(DuplicateCheckMixin, JsonUploadMixin):
     """
     Action to allow to upload a json. It is used as first step of an import.
     """
@@ -70,44 +64,10 @@ class TopicJsonUpload(DuplicateCheckMixin, Action):
 
         # validate and check for duplicates
         self.init_duplicate_set(instance["meeting_id"])
-        self.rows = [self.validate_entry(entry) for entry in data]
+        rows = [self.validate_entry(entry) for entry in data]
 
-        # generate statistics
-        itemCount, itemNew, itemError = len(self.rows), 0, 0
-        for entry in self.rows:
-            if entry["status"] == ImportStatus.CREATE:
-                itemNew += 1
-            if entry["status"] == ImportStatus.ERROR:
-                itemError += 1
-        self.statistics = {
-            "total": itemCount,
-            "created": itemNew,
-            "omitted": itemError,
-        }
-
-        # store rows in the action_worker
-        self.new_store_id = self.datastore.reserve_id(collection="action_worker")
-        fqid = fqid_from_collection_and_id("action_worker", self.new_store_id)
-        time_created = int(time())
-        self.datastore.write_action_worker(
-            WriteRequest(
-                events=[
-                    Event(
-                        type=EventType.Create,
-                        fqid=fqid,
-                        fields={
-                            "id": self.new_store_id,
-                            "result": {"import": "topic", "rows": self.rows},
-                            "created": time_created,
-                            "timestamp": time_created,
-                            "state": "running",
-                        },
-                    )
-                ],
-                user_id=self.user_id,
-                locked_fields={},
-            )
-        )
+        self.init_rows(rows)
+        self.store_rows_in_the_action_worker("topic")
         return {}
 
     def validate_entry(self, entry: Dict[str, Any]) -> Dict[str, Any]:
@@ -123,19 +83,3 @@ class TopicJsonUpload(DuplicateCheckMixin, Action):
             status = ImportStatus.ERROR
             error.append(exception.message)
         return {"status": status, "error": error, "data": entry}
-
-    def handle_relation_updates(self, instance: Dict[str, Any]) -> Any:
-        return {}
-
-    def create_events(self, instance: Dict[str, Any]) -> Any:
-        return []
-
-    def create_action_result_element(
-        self, instance: Dict[str, Any]
-    ) -> Optional[ActionResultElement]:
-        return {
-            "id": self.new_store_id,
-            "headers": self.headers,
-            "rows": self.rows,
-            "statistics": self.statistics,
-        }
