+++ openslides_backend/action/mixins/import_mixins.py
@@ -1,11 +1,12 @@
 from enum import Enum
-from typing import Any, Callable, Dict
+from time import time
+from typing import Any, Callable, Dict, List, Optional
 
 from ...shared.interfaces.event import Event, EventType
 from ...shared.interfaces.write_request import WriteRequest
 from ...shared.patterns import fqid_from_collection_and_id
 from ..action import Action
-from ..util.typing import ActionData
+from ..util.typing import ActionData, ActionResultElement
 
 
 class ImportStatus(str, Enum):
@@ -45,3 +46,66 @@ class ImportMixin(Action):
                 )
 
         return on_success
+
+
+class JsonUploadMixin(Action):
+    headers: Any
+
+    def init_rows(self, rows: List[Dict[str, Any]]) -> None:
+        self.rows = rows
+
+        # generate statistics
+        itemCount, itemCreate, itemUpdate, itemError = len(self.rows), 0, 0, 0
+        for entry in self.rows:
+            if entry["status"] == ImportStatus.CREATE:
+                itemCreate += 1
+            elif entry["status"] == ImportStatus.UPDATE:
+                itemUpdate += 1
+            elif entry["status"] == ImportStatus.ERROR:
+                itemError += 1
+        self.statistics = {
+            "total": itemCount,
+            "created": itemCreate,
+            "updated": itemUpdate,
+            "omitted": itemError,
+        }
+
+    def store_rows_in_the_action_worker(self, import_name: str) -> None:
+        self.new_store_id = self.datastore.reserve_id(collection="action_worker")
+        fqid = fqid_from_collection_and_id("action_worker", self.new_store_id)
+        time_created = int(time())
+        self.datastore.write_action_worker(
+            WriteRequest(
+                events=[
+                    Event(
+                        type=EventType.Create,
+                        fqid=fqid,
+                        fields={
+                            "id": self.new_store_id,
+                            "result": {"import": import_name, "rows": self.rows},
+                            "created": time_created,
+                            "timestamp": time_created,
+                            "state": "running",
+                        },
+                    )
+                ],
+                user_id=self.user_id,
+                locked_fields={},
+            )
+        )
+
+    def handle_relation_updates(self, instance: Dict[str, Any]) -> Any:
+        return {}
+
+    def create_events(self, instance: Dict[str, Any]) -> Any:
+        return []
+
+    def create_action_result_element(
+        self, instance: Dict[str, Any]
+    ) -> Optional[ActionResultElement]:
+        return {
+            "id": self.new_store_id,
+            "headers": self.headers,
+            "rows": self.rows,
+            "statistics": self.statistics,
+        }
