+++ tests/system/action/topic/test_json_upload.py
@@ -104,7 +104,7 @@ class TopicJsonUpload(BaseActionTestCase):
                     "data": {"title": "test", "meeting_id": 22},
                 }
             ],
-            "statistics": {"total": 1, "created": 1, "omitted": 0},
+            "statistics": {"total": 1, "created": 1, "updated": 0, "omitted": 0},
         }
 
     def test_json_upload_duplicate_in_db(self) -> None:
